xml-rpc: tolerate broken connections for SSL
With the SSL wrapper, we cannot detect a broken (TCP) connection until
we wait for the HTTP status response line. Sadly, we have to repeat the
operation that late, assuming that the request was never carried out
in the first place (there is a rare change that it would have).