[ctx_logging] a class for injecting context into loggers
Ever since py v3.7 we have context vars, that can transparently
define context for call stacks. This is particularly useful for
logs, where we want to trace that context on our records.

This class allows to set a Filter on any logger/handler, storing
the context value on any log record.